1.添加bootloader子模块 2.增加固件升级功能 3.修改flash读写接口 4.添加编译脚步

+++ shuqee_ltc_stm32f107/shuqee_ltc/Inc/user_flash.h
@@ -2,9 +2,10 @@
 #define __USER_FLASH_H
 
 #include "stm32f1xx_hal.h"
+#include "flash_map.h"
 
-extern void user_flash_read(uint16_t addr, uint16_t len, uint8_t *data);
-extern void user_flash_write(uint16_t addr, uint16_t len, uint8_t *data);
+extern void user_flash_read(uint32_t addr, uint16_t len, uint8_t *data);
+extern void user_flash_write(uint32_t addr, uint16_t len, uint8_t *data);
 extern void user_read_uuid(uint8_t *str);
 
 #endif /* __USER_FLASH_H */
